Fix file problems caused by Windows-Linux impedance for Git
Correct the executable bits on some files.  I'm not sure yet whether
this was caused by a misconfiguration in Git or simple carelessness on
my part, so additional follow-up may be needed to prevent it from
reoccurring.

Also discover that my Git configuration is sometimes converting line
endings to CRLF even though I want to use LF line endings everywhere.
(I'm working in a mix of WSL and native Windows tools, and I must have
missed some configuration on the Windows side of the house.)  In
addition to solving my problem locally, add a `.gitattributes` file to
make sure that I don't encounter the problem again on a future machine.
Note that since this project is specific to ZSH, there is not much harm
in a blanket LF-everywhere rule here.