Invert variable controlling display of Git info in prompt
Change the variable `PROMPT_GIT` (which shows Git info in the prompt
when set) to `PROMPT_HIDE_GIT` (which shows Git info in the prompt when
unset).  The net effect of this is simply to change the default state
for fresh terminal sessions.

Since the variable is more-or-less encapsulated by `toggle-git-prompt`,
the same behavior could be implemented more straightforwardly by setting
`PROMPT_GIT` directly in `prompt.zsh`, but I felt that formulating it as
a "suppression" variable rather than an "enablement" variable better
reflects the way in which I think about it, and may improve legibility a
little.