fix: Update prompt to be OS-agnostic
Some of the commands we use to build the prompt (in particular, the
`sed` and `date` calls) were written with expanded option flags for
clarity (e.g., `--silent` instead of the equivalent `-n`).
Unfortunately, the versions of these commands available by default on
Mac OS do not support the expanded form.  Revert to the short forms
(which do work identically on both Mac OS and Ubuntu/WSL).